+++ CHANGELOG.md
@@ -22,6 +22,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 - Trailing comma is now added when a single parameter is formatted onto its own line (https://github.com/facebook/ktfmt/issues/461) 
 - Migrated from `com.facebook.ktfmt` package to `org.jetbrains.ktfmt`
+- Sort declaration modifiers according to Kotlin conventions while preserving unsupported modifier
+  positions (https://github.com/facebook/ktfmt/issues/293)
 
 ### Fixed